dashboard/requirements.txt
streamlit
google-cloud-bigquery
pandas
requests
matplotlib
db-dtypes